Aerial shield stun multiplier has been returned

+++ src/fighter/common/common_param.rs
@@ -3,7 +3,7 @@
 // Shield Params
 
 pub const guard_off_attack_cancel_frame : i32 = 5;
-pub const guard_off_invalid_capture_frame_add : f32 = 3.0;
+pub const guard_off_invalid_capture_frame_add : i32 = 3;
 
 // Command Input Params
